+++ CHANGELOG.md
@@ -28,3 +28,4 @@
 - Tuple3 + Tests
 - ComputingCache + Tests
 - `org.jheaps` dependency
+- Experimental annotation